XC-294: Update Candid file

+++ canister/src/rpc_client/json/mod.rs
@@ -75,14 +75,14 @@ pub struct SendTransactionConfig {
     /// When true, skip the preflight transaction checks. Default: false.
     #[serde(rename = "skipPreflight")]
     pub skip_preflight: Option<bool>,
-    #[serde(rename = "preflightCommitment")]
     /// Commitment level to use for preflight. See Configuring State Commitment. Default finalized.
+    #[serde(rename = "preflightCommitment")]
     pub preflight_commitment: Option<String>,
     /// Maximum number of times for the RPC node to retry sending the transaction to the leader.
     /// If this parameter not provided, the RPC node will retry the transaction until it is
     /// finalized or until the blockhash expires.
     #[serde(rename = "maxRetries")]
-    pub max_retries: Option<usize>, // TODO XC-294: Is this the right type here?
+    pub max_retries: Option<u32>,
     /// Set the minimum slot at which to perform preflight transaction checks
     #[serde(rename = "minContextSlot")]
     pub min_context_slot: Option<u64>,
